requirements.txt
requests
python-dotenv==1.0.0
aiohttp==3.9.1
Django>=4,<5
djangorestframework
fastapi[all]
psycopg2-binary
sqlalchemy
asyncpg
alembic
sphinx
sphinx-autobuild
sphinx_rtd_theme
python-dotenv
asyncpg